# docs/openapi-3.1.0-locked.json
{
  "components": {
    "schemas": {
...
      "BundleBuildRequest": {
        "properties": {
          "avg_speed_kmh": {
            "default": 90.0,
            "title": "Avg Speed Kmh",
            "type": "number"
          },
          "buffer_m": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Buffer M"
          },
          "departure_iso": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Departure Iso"
          },
          "geometry": {
            "title": "Geometry",
            "type": "string"
          },
          "max_edges": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Edges"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "profile": {
            "default": "drive",
            "title": "Profile",
            "type": "string"
          },
          "route_key": {
            "title": "Route Key",
            "type": "string"
          },
          "styles": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Styles",
            "type": "array"
          },
          "trip_prefs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TripPreferences"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "plan_id",
          "route_key",
          "geometry"
        ],
        "title": "BundleBuildRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OfflineBundleManifest": {
        "properties": {
          "aqi_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aqi Key"
          },
          "aqi_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Aqi Status",
            "type": "string"
          },
          "bushfire_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bushfire Key"
          },
          "bushfire_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Bushfire Status",
            "type": "string"
          },
          "bytes_total": {
            "default": 0,
            "title": "Bytes Total",
            "type": "integer"
          },
          "cameras_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cameras Key"
          },
          "cameras_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Cameras Status",
            "type": "string"
          },
          "corridor_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corridor Key"
          },
          "corridor_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Corridor Status",
            "type": "string"
          },
          "coverage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Key"
          },
          "coverage_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Coverage Status",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "elevation_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Elevation Status",
            "type": "string"
          },
          "emergency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Emergency Key"
          },
          "emergency_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Emergency Status",
            "type": "string"
          },
          "flood_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flood Key"
          },
          "flood_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Flood Status",
            "type": "string"
          },
          "fuel_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fuel Key"
          },
          "fuel_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Fuel Status",
            "type": "string"
          },
          "hazards_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hazards Key"
          },
          "hazards_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Hazards Status",
            "type": "string"
          },
          "heritage_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heritage Key"
          },
          "heritage_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Heritage Status",
            "type": "string"
          },
          "navpack_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Navpack Status",
            "type": "string"
          },
          "places_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Places Key"
          },
          "places_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Places Status",
            "type": "string"
          },
          "plan_id": {
            "title": "Plan Id",
            "type": "string"
          },
          "rest_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rest Key"
          },
          "rest_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Rest Status",
            "type": "string"
          },
          "roadkill_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roadkill Key"
          },
          "roadkill_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Roadkill Status",
            "type": "string"
          },
          "route_key": {
            "title": "Route Key",
            "type": "string"
          },
          "school_zones_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Zones Key"
          },
          "school_zones_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "School Zones Status",
            "type": "string"
          },
          "score_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Key"
          },
          "score_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Score Status",
            "type": "string"
          },
          "styles": {
            "items": {
              "type": "string"
            },
            "title": "Styles",
            "type": "array"
          },
          "tiles_id": {
            "default": "australia",
            "title": "Tiles Id",
            "type": "string"
          },
          "toilets_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Toilets Key"
          },
          "toilets_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Toilets Status",
            "type": "string"
          },
          "traffic_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Traffic Key"
          },
          "traffic_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Traffic Status",
            "type": "string"
          },
          "weather_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weather Key"
          },
          "weather_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Weather Status",
            "type": "string"
          },
          "wildlife_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wildlife Key"
          },
          "wildlife_status": {
            "default": "missing",
            "enum": [
              "missing",
              "ready",
              "error"
            ],
            "title": "Wildlife Status",
            "type": "string"
          }
        },
        "required": [
          "plan_id",
          "route_key",
          "created_at"
        ],
        "title": "OfflineBundleManifest",
        "type": "object"
      },
...
      "RouteIntelligenceScore": {
        "properties": {
          "conditions": {
            "$ref": "#/components/schemas/RouteScoreCategory"
          },
          "data_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Data Warnings",
            "type": "array"
          },
          "overall": {
            "title": "Overall",
            "type": "number"
          },
          "overall_label": {
            "title": "Overall Label",
            "type": "string"
          },
          "safety": {
            "$ref": "#/components/schemas/RouteScoreCategory"
          },
          "services": {
            "$ref": "#/components/schemas/RouteScoreCategory"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "weather": {
            "$ref": "#/components/schemas/RouteScoreCategory"
          }
        },
        "required": [
          "overall",
          "overall_label",
          "summary",
          "safety",
          "conditions",
          "services",
          "weather"
        ],
        "title": "RouteIntelligenceScore",
        "type": "object"
      },
...
      "ScoreRefreshRequest": {
        "properties": {
          "bbox": {
            "$ref": "#/components/schemas/BBox4"
          },
          "route_key": {
            "title": "Route Key",
            "type": "string"
          }
        },
        "required": [
          "route_key",
          "bbox"
        ],
        "title": "ScoreRefreshRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Glovebox Backend",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/bundle/build": {
      "post": {
        "operationId": "build_bundle_bundle_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BundleBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfflineBundleManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Build Bundle"
      }
    },
    "/bundle/score/refresh": {
      "post": {
        "description": "Re-fetch traffic & hazards, recompute and cache the route intelligence score.",
        "operationId": "refresh_score_bundle_score_refresh_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreRefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteIntelligenceScore"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Score"
      }
    },
    "/bundle/{plan_id}": {
      "get": {
        "operationId": "get_bundle_bundle__plan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfflineBundleManifest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bundle"
      }
    },
    "/bundle/{plan_id}/download": {
      "get": {
        "operationId": "download_bundle_bundle__plan_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "plan_id",
            "required": true,
            "schema": {
              "title": "Plan Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/zip": {}
            },
            "description": "Offline trip bundle zip (manifest + tiles + corridor + places + fuel)"
          },
          "404": {
            "description": "Bundle not found for plan_id"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Bundle"
      }
    },
...
  }
}